Adapt CoreDNS dashboard to new metrics
As per documentation, the data comes now from the metric:
`coredns_proxy_request_duration_seconds_count{proxy_name="forward"}`:

CoreDNS has esentially taken advantage of the 1:1 relationship between
requests and responses to create a single metric that measures both. Each
sample in the new metric is a request and a response. The request
destination and the response code are added as labels (`to` and `rcode`,
respectively). The replacement for the old
`coredns_forward_requests_total{to}` metric is described in the
documentation as the `sum(coredns_proxy_..._{to, proxy_name="forward"})`.
In other words, the number of requests with destination `to` is the sum
of responses from `to`, regardless of their `rcode`.

The dashboard queries that need adaptation already use the sum to
aggregate by the `to` label. Since the sum is an associative operation,
we can use this sum to count requests. As an example, imagine we have the
following metrics with their corresponding labels at both times `t` and
`t - offset`. If we take the first adapted query in the change as an
example, it calculates the difference in the number of requests within a
time interval:

{rcode="NOERROR",  to="1.2.3.4"} = 100 @ t
{rcode="NXDOMAIN", to="1.2.3.4"} = 200 @ t
{rcode="SERVFAIL", to="6.7.8.9"} = 50  @ t

{rcode="NOERROR",  to="1.2.3.4"} = 10  @ t - offset
{rcode="NXDOMAIN", to="1.2.3.4"} = 10  @ t - offset
{rcode="SERVFAIL", to="6.7.8.9"} = 3   @ t - offset

The strategy followed by this commit calculates, first, the difference
between samples with the same label values:

{rcode="NOERROR",  to="1.2.3.4"} = 100 - 10 = 90
{rcode="NXDOMAIN", to="1.2.3.4"} = 200 - 10 = 190
{rcode="SERVFAIL", to="6.7.8.9"} = 50  - 3  = 47

and, second, the sum by the `to` label:

{to="1.2.3.4"} = 90 + 190 = 280
{to="6.7.8.9"}            = 47

Alternatively, if we would want to follow the documentation closely and
do the sum first, then the sum by the `to` label at times `t` and
`t - offset` would be the first step:

{to="1.2.3.4"} = 100 + 200 = 300       @ t
{to="6.7.8.9"}             = 50        @ t

{to="1.2.3.4"} = 10  + 10  = 20        @ t - offset
{to="6.7.8.9"}             = 3         @ t - offset

And then difference calculation:

{to="1.2.3.4"} = 300 - 20 = 280
{to="6.7.8.9"} = 50  - 3  = 47

Note the former approach is preferrable because it requires only one
aggregation (the latter requires two) which keeps the final query simpler.