Remove use of `externalLabels.landscape` in alert from aggregate Prometheus
This label is only available in the garden Prometheus. Following is an
example of how this alert would look until now:

```
Node ip-10-242-23-51.eu-central-1.compute.internal in landscape was not
healthy for ten scrapes in the past 30 mins.
```

Unfortunately, there is no easy way to know the landscape based on the labels.