Deploy KSM to seed if it is garden cluster
Previously, the gardenlet did not deploy the kube-state-metrics instance to the
garden namespace of the seed, if the seed cluster was a garden cluster. This was
done to avoid having two kube-state-metrics instances in the garden namespace:
one deployed by the gardener operator and the other deployed by the gardenlet,
because they used to be equivalent.

The two kube-state-metrics instances are going to be different in the near
future. Hence, both should be deployed to the garden namespace if the seed
cluster is the Garden cluster.

- The kube-state-metrics instance deployed by the gardener operator is going to
  be used to monitor the Garden resource in the runtime-garden cluster in a
  future PR, in addition to the pods in the runtime-garden cluster.

- The kube-state-metrics instance deployed by the gardenlet is going to be used
  to monitor the pods in the seed cluster, as before.

So although there is an overlap in the 2 kube-state-metrics instances (both
monitor the pods), only the runtime instance is going to monitor the Garden
resource.

In the local setup, the garden cluster is a seed cluster, so by deploying both
instances we can try and test both configurations.

Note that in real landscapes, the Garden cluster is typically not used as a seed
cluster.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>