Re-introduce the kube_pod_labels metric
The upgrade to kube-state-metrics v2.12 disables the kube_pod_labels
metric by default and needs to be re-enabled explicitly. Only the
origin label is required by Gardener.

This metric is used in the following two alerting rules:

- alert: KubePodPendingShoot
  expr: (kube_pod_status_phase{phase="Pending", type="shoot"} == 1 and on(pod) kube_pod_labels{label_origin="gardener"})
- alert: KubePodNotReadyShoot
  expr: (kube_pod_status_ready{condition="true", type="shoot"} == 0 and on(pod) kube_pod_labels{label_origin="gardener"})

This was broken since the upgrade to kube-state-metrics 2.5.0 where the
kube_pod_labels metric existed but did not have an origin label, unless
explicitly configured.