Server-side filter for allowed kube-state-metrics
kube-state-metrics can be configured to expose only specific metrics
about the state of objects. In Gardener, we are not interested in
ingesting all available metrics into Prometheus but a subset of them.
Previously, this filtering was achieved via the scrape configuration
on the Prometheus side.

The kube-state-metrics in the seed was already configured to expose
certain metrics, but the kube-state-metrics in the shoot was not. Thus,
when the shoot Prometheus scraped the latter kube-state-metrics, it
downloaded a bunch of unwanted metrics and filtered them out via metric
relabeling. This is inefficient because such filtering can be done
server-side to reduce the scrape time.

This commit improves this by extending the kube-state-metrics
configurations to make sure server-side filtering is in-place. This is
done via the `--metric-allowlist` flag, which expects a comma-separated
list of regular expressions. Metric names that match at least one of those
regular expressions are exposed. Therefore, metric names in
`--metric-allowlist` are now wrapped around the ^ and $ characters to fully
match the metric name.

The client-side filtering from the Prometheus scrape configuration is now
removed as it is not necessary. The kube-state-metrics deployments expose
only the desired metrics and Prometheus injects everything that is scraped.
Keeping the client-side filtering could be seen as a second line of defense,
but it would also add a bit of configuration overhead without a clear
benefit.