Wait for previous ManagedResource to be healthy
The cluster role and cluster role binding were originally managed by the
shoot-core-kube-state-metrics ManagedResource but are now managed by the
shoot-core-kube-state-metrics-target ManagedResource.

In order to migrate those two resources to the latter ManagedResource,
we need to wait until the former stops managing them and deletes them.
Otherwise, the previous ManagedResource wants to delete them while the
new one keeps creating them.

As for the tests, a second goroutine is introduced to listen to the
shoot-core-kube-state-metrics ManagedResource and set its generation and
status conditions to healthy.