Add alert for capped VPA recommendations for the seed components
This alert needs to be deployed within the aggregate Prometheus because
this one is connected with the seed Alertmanager. Also, the garden
Prometheus federates from it so this is a way to have these alerts also
available in the garden Prometheus.

We provide a new suffix "Shoot" or "Seed" in the rule name to better
differentiate between the capped recommendation from the shoot or the
seed components.

There is only one aggregate Prometheus also in case the garden cluster
is a seed, so we do not need to add the addition kube-state-metrics
suffix.