Fix Prometheus rule tests in etcd
The unit tests were originally misconfigured because they query a test
time series that only exists for 10 minutes and gets staled for 5
additional minutes. Note such time series is defined as 0x20 but the
evaluation interval is 30 seconds. Furthermore, the tests would check
the alert after 15 minutes despite it triggers after 10 minutes. It
seems, in this situation, the alert remains triggered after 15 minutes
in Prometheus v2 but not in Prometheus v3.

This commit adjusts the tests to evaluate the alert after 10 minutes, in
line with its configuration. This resolves the issue in Prometheus v3
where the alert does not remain triggered after 15 minutes.