Add network policy to allow federating from the cache Prometheus
The metering Prometheus only ingests data from the cache Prometheus.
Therefore, allowing connections to the cache Prometheus should suffice
for federation to work. However, the metering Prometheus pods are
labeled for network policies that allow access to any scrape targets
in the seed. This is done so to keep it consistent with the other
Prometheus instances in the seed.