Proposal: Use separate struct to configure dashboards
This commit is simply an idea / proposal for a different but yet similar
way to tackle this problem. I found a similar problem happens for the
Istio dashboards, where they are always deployed for the seed Plutono
independently of the value passed in the function argument.

The proposal is to move the dashboard configuration booleans into their
own struct so that a default is set if not specified, e.g., for the
seed Plutono:

	return sharedcomponent.NewPlutono(
		...
		plutono.DashboardValues{
			ExcludeVPAInstallationDashboards: !v1beta1helper.SeedSettingVerticalPodAutoscalerEnabled(seed.GetInfo().Spec.Settings),
		},
	)

This way, the code becomes more readable because only those parameters
relevant for each specific component need to be set, e.g., in the seed
example above, we don't need to set isWorkerless anymore because it only
matters for shoots.

The DashboardValues struct becomes some sort of exclusion list. The
default behaviour is to include everything and exclude specific
dashboards. This has two benefits. First, that if we ever forget to set
some exclusion, the dashboard will still be deployed despite probably
broken. I figured this is better than missing a dashboard completely,
which would be the case if we forget to include it. Second, dashboards
can only be removed but not added, e.g., we can't include the seed
dashboards into a shoot Plutono.

The second significant change is that the VPA installation dashboards
can also be deployed independently of the common dashboards, similar as
for the VPA custom resources. With this, it is now possible to deploy
the VPA installation dashboards in the runtime cluster, e.g., if the VPA
is installed by an external component. Anyhow, we could probably discuss
moving the VPA dashboard folders outside common. If they are not always
deployed, then they are not common dashboards.

Finally, I adapted some of the unit tests, but I don't like them because
they are just counting how many dashboard are deployed. I think we could
find better tests for this.