Prepare migration to remove Prometheus obsolete folders
This migration is thought for shoot Prometheus instances where it can't
be guaranteed the init container introduced in previous commit will run
after successful reconciliation, e.g., shoots with failed reconciliation
or hibernated shoots.

This commit only prepares to execute this migration as part of the
startup of the gardenlet. It iterates over the Cluster resources in the
seed to find shoots, but nothing is yet executed.