Ignore the shoot Prometheus ManagedResource before patching
Shoot Prometheis are actually managed by the GRM. This means any patches
applied in the following commits will be instantly reverted, unless the
corresponding ManagedResource is ignored via annotation:

    resources.gardener.cloud/ignore="true"

The ignore annotation can be removed once the migration is complete.
This has the benefit that any patches applied will be undone and the
shoot brought back to its desired state by the GRM. We choose to
unignore the ManagedResource in a deferred call to make sure the
annotation is always removed, also in case of failures.