Prepare plutono unit tests to validate list of dashboards
A list of dashboards file names can be passed as a second argument to
validate the actual list instead of the number of dashboards. If this
list is nil, tests fall back to validating the number of dashboards.
This is an intermediate step until all unit tests have been adapted.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>