Allow network access from GRM when runtime cluster is seed
When the runtime cluster is also a seed, the gardenlet does not deploy a
new GRM because the GRM deployed by the Garden operator suffices. In
such situations, such runtime GRM needs network access to additional
Prometheus to check for their health.

This network access is allowed always, but it has no effect when the
runtime cluster is not a seed.