Rework service configuration for the longterm Prometheus
This commit exposes the regular port 9090 of the longterm Prometheus on
port 80 of the service. The cortex container is now exposed on port 81,
with 9091 as target port.

This aligns the longterm Prometheus service with other
Prometheus services and adds an additional service port for Cortex.
Before, only the Cortex container was exposed.

With this change, the network controller will create network policies to
access the longterm Prometheus on port 9090, exactly as it currently
does for other Prometheus.

The longterm Prometheus ingress has been changed to keep mapping to the
Cortex frontend, now on the service port 81 as mentioned above.

TODO: This commit needs split