Allow network access from the gardenlet to Prometheus
This allow access from the gardenlet to the aggregate, cache and seed
Prometheus instances in the garden namespace in the seed, and to all
shoot Prometheus instances in their corresponding shoot control plane
namespace.

After this change, the gardenlet is able to query those Prometheus
instances and perform more complex health checks such as query for
active alerts.