Introduce new seed condition for observability components
The GardenletReady and SeedSystemComponentsHealthy conditions have the
potential to block shoot operations if they become unhealthy. This makes
sense to prevent shoots from being operated while the corresponding seed
is unhealthy.

Until now, health of observability components such as Prometheus,
Plutono or Vali were monitored in the SeedSystemComponentsHealthy
condition. In consequence, the observability components could block
shoot operations from happening, despite not being critical. This
becomes specially problematic when extending observability health checks
to consider active alerts.

To circumvent this, this commit introduces a new
SeedObservabilityComponentsHealthy condition specific for the
observability components in the seed. Follow-up commits will start using
it. The "Seed" prefix is required to distinguish it from the regular
ObservabilityComponentsHealthy condition in the shoot and garden
clusters in cases where the seed is also a shoot or a garden cluster,
respectively.